backend/requirements.txt
# GeoHIS Backend Dependencies
# Production-ready configuration

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1

# Configuration
python-decouple==3.8

# HTTP Client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging & Monitoring
structlog==23.2.0

# Security
email-validator==2.1.0

# =============================================================================
# Scientific Computing & Analysis
# =============================================================================

# Core Scientific - PINNED for stability
# numpy must be <2.0 for compatibility with bottleneck, rasterio, and other C extensions
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<1.14.0
pandas>=2.0.0,<2.2.0

# Machine Learning
scikit-learn>=1.3.0,<1.5.0
xgboost>=2.0.0,<2.1.0

# Geospatial Analysis
shapely>=2.0.0,<2.1.0
pyproj>=3.6.0,<3.7.0
rasterio>=1.3.0,<1.4.0

# Visualization
matplotlib>=3.7.0,<3.9.0
seaborn>=0.13.0,<0.14.0